Authorize model before destroy

+++ app/controllers/krudmin/application_controller.rb
@@ -150,6 +150,8 @@ module Krudmin
     end
 
     def destroy
+      authorize model
+
       respond_to do |format|
         if model.destroy
           format.html { redirect_to resource_root, notice: destroyed_message }
